Initial commit
cdj_mode is meant to be sourced into a shell. It defines functions that
can be used to prepare a track (bpm, cue points, wav conversion) for
DJ'ing, and then copy that track to multiple target directories
simultaneously (typically, each target directory is a usb key that will
be inserted into a CDJ).